Added a new convolution algorthm: "simple". It's just a naive triple-loop sum of the products of corresponding entries in the filter and data.

Renamed the GEMM-accelerated convolution algorithm to "unrolled".

Refactored convolution nodes somewhat. Now "ConvolutionalLayerNode" refines itself into a SimpleConvolutionNode, an UnrolledConvolutionNode, or a DiagonalConvolutionNode.

Cleaned up `ReceptiveFieldMatrixNode` implementation.

A couple of miscellaneous small items crept in during the above work:
* Added no-argument overload of `IRFunctionEmitter::LocalScalar()` to return an uninitialized `IRLocalScalar`.
* Added assignment operator to `IRLocalScalar`.
* You no longer have to explicitly call `DeclarePrintf` if you're going to print something from emitted code. The module emitter does it for you automatically if necessary.
* Removed a couple of unnecessary functions from emitter objects

Related work items: #1301

+++ libraries/nodes/include/ReceptiveFieldMatrixNode.h
@@ -28,7 +28,7 @@ namespace ell
 namespace nodes
 {
     /// <summary>
-    /// A node that reshapes an image into a format suitable for computing convolutions via matrix multiplcation.
+    /// A node that reshapes an image into a format suitable for computing convolutions via matrix multiplication.
     /// Also known as Im2Col.
     /// </summary>
     template <typename ValueType>
@@ -56,12 +56,12 @@ namespace nodes
         /// <param name="outputHeight"> The output image height. </param>
         ReceptiveFieldMatrixNode(const model::PortElements<ValueType>& input,
                                  const model::PortMemoryLayout& inputMemoryLayout,
-                                 size_t filterWidth,
-                                 size_t stride,
-                                 size_t convolutionPadding,
+                                 int filterWidth,
+                                 int stride,
+                                 int convolutionPadding,
                                  std::array<int, 3> dataOrder, // 0, 1, 2 == rows, columns, channels -- row-major
-                                 size_t outputWidth,
-                                 size_t outputHeight);
+                                 int outputWidth,
+                                 int outputHeight);
 
         /// <summary> Gets information about the input memory layout </summary>
         const model::PortMemoryLayout& GetInputMemoryLayout() const { return _inputMemoryLayout; }
@@ -97,12 +97,12 @@ namespace nodes
         model::OutputPort<ValueType> _output;
 
         model::PortMemoryLayout _inputMemoryLayout;
-        size_t _filterWidth;
-        size_t _stride;
-        size_t _convolutionPadding;
+        int _filterWidth;
+        int _stride;
+        int _convolutionPadding;
         std::array<int, 3> _dataOrder;
-        size_t _outputWidth;
-        size_t _outputHeight;
+        int _outputWidth;
+        int _outputHeight;
     };
 }
 }
